pyproject.toml
[project]
name = "churn-x"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=2.0.7",
    "category-encoders>=2.8.1",
    "certifi>=2025.8.3",
    "dagshub>=0.6.3",
    "dill>=0.4.0",
    "dotenv>=0.9.9",
    "faker>=37.8.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "groq>=0.31.1",
    "imblearn>=0.0",
    "ipywidgets>=8.1.7",
    "joblib>=1.5.2",
    "matplotlib>=3.10.6",
    "mcp>=1.14.1",
    "mlflow>=3.3.2",
    "nbformat>=5.10.4",
    "numpy>=2.2.6",
    "optuna>=4.5.0",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pymongo>=4.15.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.43",
    "statsmodels>=0.14.5",
    "xgboost>=3.0.5",
]
